initialization and managerloginmenu addition
